[Refactor][SpecDecode] Move AscendSpecDecodeBaseProposer to llm_base_proposer.py
- Move AscendSpecDecodeBaseProposer class from eagle_proposer.py to new file llm_base_proposer.py

- Update import paths in draft_proposer.py and eagle_proposer.py

- Simplify eagle_proposer.py to only contain AscendEagleProposer class definition

- Update test imports to use new module path
