+++ vllm_ascend/spec_decode/draft_proposer.py
@@ -2,7 +2,7 @@ import torch
 from vllm.config import VllmConfig
 from vllm.v1.spec_decode.draft_model import DraftModelProposer
 
-from vllm_ascend.spec_decode.eagle_proposer import AscendSpecDecodeBaseProposer
+from vllm_ascend.spec_decode.llm_base_proposer import AscendSpecDecodeBaseProposer
 
 
 class AscendDraftModelProposer(DraftModelProposer, AscendSpecDecodeBaseProposer):
